Rename `npm-util.js` and improve method naming
