+++ source/version.js
@@ -7,7 +7,7 @@ class Version {
 	}
 
 	isPrerelease() {
-		return module.exports.PRERELEASE_VERSIONS.includes(this.version) || Boolean(semver.prerelease(this.version));
+		return Boolean(semver.prerelease(this.version));
 	}
 
 	satisfies(range) {
@@ -45,6 +45,8 @@ module.exports = version => new Version(version);
 
 module.exports.SEMVER_INCREMENTS = ['patch', 'minor', 'major', 'prepatch', 'preminor', 'premajor', 'prerelease'];
 module.exports.PRERELEASE_VERSIONS = ['prepatch', 'preminor', 'premajor', 'prerelease'];
+
+module.exports.isPrereleaseOrIncrement = input => module.exports(input).isPrerelease() || module.exports.PRERELEASE_VERSIONS.includes(input);
 
 const isValidVersion = input => Boolean(semver.valid(input));
